Move BaseTreeView and associated methods to its own file
